Move `ListSlice` tests to a separate file
